replaced TimeSpan with scala.concurrent.duration.Duration
TimeSpan has ambiguous definition. It can behave like a joda Period (e.g. take care about timeshift)
but for most cases it should by only a wrapper for in-milliseconds duration. This change replace
TimeSpan with Duration deprecating the first one.
