+++ core/util/src/test/scala/net/liftweb/util/TimeHelpersSpec.scala
@@ -19,6 +19,7 @@ package util
 
 import java.util.{Calendar, Date}
 
+import org.joda.time.DateTime
 import org.specs2.mutable.Specification
 import org.specs2.ScalaCheck
 import org.specs2.time.NoTimeConversions
@@ -35,85 +36,6 @@ import TimeHelpers._
  */
 object TimeHelpersSpec extends Specification with ScalaCheck with TimeAmountsGen with NoTimeConversions {
   "TimeHelpers Specification".title
-
-  "A TimeSpan" can {
-    "be created from a number of milliseconds" in {
-      TimeSpan(3000) must_== TimeSpan(3 * 1000)
-    }
-    "be created from a number of seconds" in {
-      3.seconds must_== TimeSpan(3 * 1000)
-    }
-    "be created from a number of minutes" in {
-      3.minutes must_== TimeSpan(3 * 60 * 1000)
-    }
-    "be created from a number of hours" in {
-      3.hours must_== TimeSpan(3 * 60 * 60 * 1000)
-    }
-    "be created from a number of days" in {
-      3.days must_== TimeSpan(3 * 24 * 60 * 60 * 1000)
-    }
-    "be created from a number of weeks" in {
-      3.weeks must_== TimeSpan(3 * 7 * 24 * 60 * 60 * 1000)
-    }
-    "be converted implicitly to a date starting from the epoch time" in {
-      3.seconds.after(new Date(0)) must beTrue
-    }
-    "be converted to a date starting from the epoch time, using the date method" in {
-      3.seconds.after(new Date(0)) must beTrue
-    }
-    "be implicitly converted to a Long" in {
-      (3.seconds == 3000L) must_== true
-    }
-    "be compared to an int" in {
-      (3.seconds == 3000) must_== true
-      (3.seconds != 2000) must_== true
-    }
-    "be compared to a long" in {
-      (3.seconds == 3000L) must_== true
-      (3.seconds != 2000L) must_== true
-    }
-    "be compared to another TimeSpan" in {
-      3.seconds must_== 3.seconds
-      3.seconds must_!= 2.seconds
-    }
-    "be compared to another object" in {
-      3.seconds must_!= "string"
-    }
-  }
-
-  "A TimeSpan" should {
-    "return a new TimeSpan representing the sum of the 2 times when added with another TimeSpan" in {
-      3.seconds + 3.seconds must_== 6.seconds
-    }
-    "return a new TimeSpan representing the difference of the 2 times when substracted with another TimeSpan" in {
-      3.seconds - 4.seconds must_== (-1).seconds
-    }
-    "have a later method returning a date relative to now plus the time span" in {
-      val expectedTime = new Date().getTime + 3.seconds.millis
-
-      3.seconds.later.getTime must beCloseTo(expectedTime, 1000L)
-    }
-    "have an ago method returning a date relative to now minus the time span" in {
-      val expectedTime = new Date().getTime - 3.seconds.millis
-
-      3.seconds.ago.getTime must beCloseTo(expectedTime, 1000L)
-    }
-    "have a toString method returning the relevant number of weeks, days, hours, minutes, seconds, millis" in {
-      val conversionIsOk = forAll(timeAmounts)((t: TimeAmounts) => { val (timeSpanToString, timeSpanAmounts) = t
-        timeSpanAmounts forall { case (amount, unit) =>
-          amount >= 1  &&
-          timeSpanToString.contains(amount.toString) || true }
-      })
-      val timeSpanStringIsPluralized = forAll(timeAmounts)((t: TimeAmounts) => { val (timeSpanToString, timeSpanAmounts) = t
-        timeSpanAmounts forall { case (amount, unit) =>
-               amount > 1  && timeSpanToString.contains(unit + "s") ||
-               amount == 1 && timeSpanToString.contains(unit) ||
-               amount == 0 && !timeSpanToString.contains(unit)
-        }
-      })
-      check(conversionIsOk && timeSpanStringIsPluralized)
-    }
-  }
 
   "the TimeHelpers" should {
     "provide a 'seconds' function transforming a number of seconds into millis" in {
@@ -136,8 +58,9 @@ object TimeHelpersSpec extends Specification with ScalaCheck with TimeAmountsGen
     }
 
     "make sure noTime does not change the day" in {
-      dateFormatter.format(0.days.ago.noTime) must_== dateFormatter.format(new Date())
-      dateFormatter.format(3.days.ago.noTime) must_== dateFormatter.format(new Date(millis - (3 * 24 * 60 * 60 * 1000)))
+      val dateWithoutTime = new DateTime(2015, 1, 1, 0, 0)
+      val dateWithTime = dateWithoutTime.withHourOfDay(2).withMinuteOfHour(3)
+      dateWithTime.toDate.noTime must_== dateWithoutTime.toDate
     }
 
     "provide a day function returning the day of month corresponding to a given date (relative to UTC)" in {
@@ -222,24 +145,3 @@ object TimeHelpersSpec extends Specification with ScalaCheck with TimeAmountsGen
     }
   }
 }
-
-
-trait TimeAmountsGen {
-
-  type TimeAmounts = (String, List[(Int, String)])
-
-  val timeAmounts =
-    for {
-      w <- choose(0, 2)
-      d <- choose(0, 6)
-      h <- choose(0, 23)
-      m <- choose(0, 59)
-      s <- choose(0, 59)
-      ml <- choose(0, 999)
-    }
-    yield (
-      TimeSpan(weeks(w) + days(d) + hours(h) + minutes(m) + seconds(s) + ml).toString,
-      (w, "week") :: (d, "day") :: (h, "hour") :: (m, "minute") :: (s, "second") :: (ml, "milli") :: Nil
-    )
-}
-
