+++ web/webkit/src/main/scala/net/liftweb/http/LiftRules.scala
@@ -29,6 +29,7 @@ import JE._
 import JsCmds._
 import auth._
 
+import scala.concurrent.duration._
 import scala.xml._
 import java.util.{Locale, TimeZone, ResourceBundle, Date}
 import java.io.{InputStream, ByteArrayOutputStream, BufferedReader, StringReader}
@@ -240,7 +241,7 @@ class LiftRules() extends Factory with FormVendor with LazyLoggable {
    * for general notices (not associated with id-s) regardless if they are set for the page rendering, ajax
    * response or Comet response.
    */
-  val noticesAutoFadeOut = new FactoryMaker[(NoticeType.Value) => Box[(TimeSpan, TimeSpan)]]((notice : NoticeType.Value) => Empty){}
+  val noticesAutoFadeOut = new FactoryMaker[(NoticeType.Value) => Box[(Duration, Duration)]]((notice : NoticeType.Value) => Empty){}
 
   /**
    * Use this to apply various effects to the notices. The user function receives the NoticeType
@@ -687,7 +688,7 @@ class LiftRules() extends Factory with FormVendor with LazyLoggable {
    * the long polling connection
    */
   val clientActorLifespan = new FactoryMaker[LiftActor => Long](
-    () => (actor: LiftActor) => (30.minutes): Long
+    () => (actor: LiftActor) => 30.minutes.toMillis
   ){}
 
   /**
@@ -710,17 +711,17 @@ class LiftRules() extends Factory with FormVendor with LazyLoggable {
   /**
    * If a Comet request fails timeout for this period of time. Default value is 10 seconds
    */
-  @volatile var cometFailureRetryTimeout: Long = 10.seconds
+  @volatile var cometFailureRetryTimeout: Long = 10.seconds.toMillis
 
   /**
    * The timeout in milliseconds of a comet ajax-request. Defaults to 5000 ms.
    */
-  @volatile var cometProcessingTimeout: Long = 5.seconds
+  @volatile var cometProcessingTimeout: Long = 5.seconds.toMillis
 
   /**
    * The timeout in milliseconds of a comet render-request. Defaults to 30000 ms.
    */
-  @volatile var cometRenderTimeout: Long = 30.seconds
+  @volatile var cometRenderTimeout: Long = 30.seconds.toMillis
 
   /**
    * The dispatcher that takes a Snippet and converts it to a
@@ -1042,7 +1043,7 @@ class LiftRules() extends Factory with FormVendor with LazyLoggable {
   /**
    * How long should we wait for all the lazy snippets to render
    */
-  val lazySnippetTimeout: FactoryMaker[TimeSpan] = new FactoryMaker(() => 30.seconds) {}
+  val lazySnippetTimeout: FactoryMaker[Duration] = new FactoryMaker(() => 30.seconds: Duration) {}
 
   /**
    * Does the current context support parallel snippet execution
@@ -1616,13 +1617,13 @@ class LiftRules() extends Factory with FormVendor with LazyLoggable {
    * (given in milliseconds) will be discarded, hence eligible for garbage collection.
    * The default value is 10 minutes.
    */
-  @volatile var unusedFunctionsLifeTime: Long = 10.minutes
+  @volatile var unusedFunctionsLifeTime: Long = 10.minutes.toMillis
 
   /**
    * The polling interval for background Ajax requests to prevent functions of being garbage collected.
    * Default value is set to 75 seconds.
    */
-  @volatile var liftGCPollingInterval: Long = 75.seconds
+  @volatile var liftGCPollingInterval: Long = 75.seconds.toMillis
 
   /**
    * Put a test for being logged in into this function
@@ -1633,7 +1634,7 @@ class LiftRules() extends Factory with FormVendor with LazyLoggable {
    * The polling interval for background Ajax requests to keep functions to not be garbage collected.
    * This will be applied if the Ajax request will fail. Default value is set to 15 seconds.
    */
-  @volatile var liftGCFailureRetryTimeout: Long = 15.seconds
+  @volatile var liftGCFailureRetryTimeout: Long = 15.seconds.toMillis
 
   /**
    * If this is Full, comet updates (partialUpdates or reRenders) are
