+++ web/webkit/src/main/scala/net/liftweb/http/js/jquery/JqJsCmds.scala
@@ -16,12 +16,14 @@
 package net.liftweb 
 package http 
 package js 
-package jquery 
+package jquery
 
+import net.liftweb.util.TimeSpanHelpers.TimeSpan
+
+import scala.concurrent.duration._
 import scala.xml.NodeSeq
 import net.liftweb.util.Helpers._
 import net.liftweb.util.Helpers
-import net.liftweb.util.TimeHelpers
 import net.liftweb.common._
 import net.liftweb.util._
 
@@ -510,7 +512,17 @@ object JqJsCmds {
      * @param uid the element id
      * @param time the duration of the effect.
      */
-    def apply(uid: String, time: TimeSpan) = new Show(uid, Full(time))
+    @deprecated("Use scala.concurrent.duration.Duration instead of TimeSpan")
+    def apply(uid: String, time: TimeSpan) = new Show(uid, Full(time.toScalaDuration))
+
+    /**
+     *
+     * Show an element identified by uid
+     *
+     * @param uid the element id
+     * @param time the duration of the effect.
+     */
+    def apply(uid: String, time: Duration) = new Show(uid, Full(time))
   }
 
   /**
@@ -519,7 +531,7 @@ object JqJsCmds {
    * @param uid the element id
    * @param time the duration of the effect.
    */
-  class Show(val uid: String, val time: Box[TimeSpan]) extends JsCmd with HasTime {
+  class Show(val uid: String, val time: Box[Duration]) extends JsCmd with HasTime {
     def toJsCmd = "try{jQuery(" + ("#" + uid).encJs + ").show(" + timeStr + ");} catch (e) {}"
   }
 
@@ -540,20 +552,39 @@ object JqJsCmds {
      * @param uid the element id
      * @param time the duration of the effect.
      */
-    def apply(uid: String, time: TimeSpan) = new Hide(uid, Full(time))
+    @deprecated("Use scala.concurrent.duration.Duration instead of TimeSpan")
+    def apply(uid: String, time: TimeSpan) = new Hide(uid, Full(time.toScalaDuration))
+
+    /**
+     * Hide an element identified by uid
+     *
+     * @param uid the element id
+     * @param time the duration of the effect.
+     */
+    def apply(uid: String, time: Duration) = new Hide(uid, Full(time))
   }
 
   /**
    * Hide an element identified by uid and the animation will last @time
    */
-  class Hide(val uid: String, val time: Box[TimeSpan]) extends JsCmd with HasTime {
+  class Hide(val uid: String, val time: Box[Duration]) extends JsCmd with HasTime {
     def toJsCmd = "try{jQuery(" + ("#" + uid).encJs + ").hide(" + timeStr + ");} catch (e) {}"
   }
 
   /**
    * Show a message msg in the element with id where for duration milliseconds and fade out in fadeout milliseconds
    */
-  case class DisplayMessage(where: String, msg: NodeSeq, duration: TimeSpan, fadeTime: TimeSpan) extends JsCmd {
+  @deprecated("Use scala.concurrent.duration.Duration instead of TimeSpan")
+  object DisplayMessage {
+    def apply(where: String, msg: NodeSeq, duration: TimeSpan, fadeTime: TimeSpan): DisplayMessage = {
+      new DisplayMessage(where, msg, duration.toScalaDuration, fadeTime.toScalaDuration)
+    }
+  }
+
+  /**
+   * Show a message msg in the element with id where for duration milliseconds and fade out in fadeout milliseconds
+   */
+  case class DisplayMessage(where: String, msg: NodeSeq, duration: Duration, fadeTime: Duration) extends JsCmd {
     def toJsCmd = (Show(where) & JqSetHtml(where, msg) & After(duration, Hide(where, fadeTime))).toJsCmd
   }
 
@@ -565,20 +596,41 @@ object JqJsCmds {
     * Fade Out with the default duration and fadeTime provided by JsRules
     */
     def apply(id: String) = new FadeOut(id, JsRules.prefadeDuration, JsRules.fadeTime)
+
+
+    /**
+     * Fades out the element having the provided id, by waiting
+     * for the given duration and fading out during fadeTime
+     */
+    @deprecated("Use scala.concurrent.duration.Duration instead of TimeSpan")
+    def apply(id: String, duration: TimeSpan, fadeTime: TimeSpan): FadeOut = {
+      new FadeOut(id, duration.toScalaDuration, fadeTime.toScalaDuration)
+    }
   }
 
   /**
    * Fades out the element having the provided id, by waiting
    * for the given duration and fading out during fadeTime
    */
-  case class FadeOut(id: String, duration: TimeSpan, fadeTime: TimeSpan) extends JsCmd {
-    def toJsCmd = (After(duration, JqJE.JqId(id) ~> (new JsRaw("fadeOut(" + fadeTime.millis + ")") with JsMember))).toJsCmd
+  case class FadeOut(id: String, duration: Duration, fadeTime: Duration) extends JsCmd {
+    def toJsCmd = (After(duration, JqJE.JqId(id) ~> (new JsRaw("fadeOut(" + fadeTime.toMillis + ")") with JsMember))).toJsCmd
   }
 
   /**
   * The companion object to FadeIn that provides an alternative factory
   */
   object FadeIn {
+
+    /**
+     * Fades in the element having the provided id, by waiting
+     * for the given duration and fading in during fadeTime
+     * and use @fadeTime
+     */
+    @deprecated("Use scala.concurrent.duration.Duration instead of TimeSpan")
+    def apply(id: String, duration: TimeSpan, fadeTime: TimeSpan): FadeIn = {
+      new FadeIn(id, duration.toScalaDuration, fadeTime.toScalaDuration)
+    }
+
     /**
     * Fade In with the default duration and fadeTime provided by JsRules
     */
@@ -590,8 +642,8 @@ object JqJsCmds {
    * for the given duration and fading in during fadeTime
    * and use @fadeTime
    */
-  case class FadeIn(id: String, duration: TimeSpan, fadeTime: TimeSpan) extends JsCmd {
-    def toJsCmd = (After(duration, JqJE.JqId(id) ~> (new JsRaw("fadeIn(" + fadeTime.millis + ")") with JsMember))).toJsCmd
+  case class FadeIn(id: String, duration: Duration, fadeTime: Duration) extends JsCmd {
+    def toJsCmd = (After(duration, JqJE.JqId(id) ~> (new JsRaw("fadeIn(" + fadeTime.toMillis + ")") with JsMember))).toJsCmd
   }
 
   /**
